+++ Sorting/knuth_gap_strategy.cpp
@@ -1,9 +1,9 @@
 #include "knuth_gap_strategy.h"
 
 
-vector<int> KnuthGapStrategy::getGaps(size_t n) const {
+vector<size_t> KnuthGapStrategy::getGaps(size_t n) const {
 	size_t k = log(2 * (n / 3) + 1) / log(3);
-	vector<int> ans(k);
+	vector<size_t> ans(k);
 
 	ans[k - 1] = 1;  // ensuring valid sorting
 	for (int i = (int)k - 2; i >= 0; i--) {
